// src/enum.ts
export enum Visibility {
    PRIVATE = 'private',
    PROTECTED = 'protected',
    PUBLIC = 'public',
    STATIC = 'static',
    ABSTRACT = 'abstract'
}

export enum ObjectType {
    PACKAGE = 'package',
    CLASS = 'class',
    ABSTRACT = 'abstract',
    ENTITY = 'class',
    ENUM = 'enum',
    ENUM_ENTRY = '',
    INTERFACE = 'interface',
    METHOD = 'method',
    PROPERTY = 'property'
}